+++ dna_fe/src/styles/admin/ordersPage.css
@@ -3,6 +3,7 @@
   background-color: #f3f4f6;
   min-height: 100vh;
   font-family: Arial, sans-serif;
+  text-align: center;
 }
 
 .orders-buttons {
